ready to work on encryption methods
